refactor package, merge "data" to "model"
