+++ app/src/main/java/com/singtel/groupit/injection/component/ApplicationComponent.java
@@ -3,7 +3,7 @@ package com.singtel.groupit.injection.component;
 import android.app.Application;
 
 
-import com.singtel.groupit.data.DataManager;
+import com.singtel.groupit.DataManager;
 import com.singtel.groupit.injection.module.ApplicationModule;
 import com.singtel.groupit.view.activity.MainActivity;
